Fix quantization for input to reference model
Add the zerpoint instead of subtracting. This worked since the tests so
far used the ones as inputs which quantize to a zp of -128 which gives
the same np.int8 result in both cases since the int8 wraps.
Also needs to round and clip the scaled values to the int8 range.

Signed-off-by: Per Åstrand <per.astrand@arm.com>

Change-Id: Ideaed6d072a4065573b38fb7476c7dbe8ba814fd